renamed cake receipt emails
