+++ src/components/BespokeCakesForm.tsx
@@ -30,7 +30,7 @@ import {
 import { Input } from '@/components/ui/input';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 
-import { receiptEmail } from '../mailer/receiptEmail';
+import { receiptEmail } from '../mailer/cakeOrderReceiptEmail';
 import { cakeOrderEmail } from '@/mailer/cakeOrderEmail';
 
 const formSchema = z.object({
